+++ packages/design-system-angular/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/geovistory/design-system/compare/v1.0.0...v1.0.1) (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
+## 1.0.0 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-angular
+
+
+
+
+
 ## 0.3.1-pr-14.8 (2022-07-07)
 
 **Note:** Version bump only for package @geovistory/design-system-angular